Updating breaking-up-markdowns.sh

+++ breaking-up-markdowns.sh
@@ -1,10 +1,14 @@
-
 function moveit {
-    git switch main
-    git checkout gitsource_main $1
-    git add $1
-    git commit -m'move to public'
-    git switch gitsource_main
-    git rm -r $1
-    git commit -m'moved to public'
+    if [ -z $1 ] ; then
+        echo need a file
+    else
+        git switch new
+        git checkout old $1
+        git add $1
+        git commit -m"move $1 to public"
+        git switch old
+        git rm -r $1
+        git commit -m"moved $1 to public"
+        git switch new
+    fi
 }
